Module egglog::ast

Modules

desugar

parse

Structs

FunctionDecl

Represents the declaration of a function directly parsed from source syntax.

Id

IdentSort

Metadata

NormCommand

NormFunctionDecl

A normalized function declaration- the desugared version of a FunctionDecl. TODO so far only the merge action is normalized, not the default value or merge expression.

NormRule

NormRunConfig

Rewrite

Rule

RunConfig

Schema

Symbol

A interned string in the global symbol table.

Variant

Enums

Action

Command

A Command is the top-level construct in egglog. It includes defining rules, declaring functions, adding to tables, and running rules (via a Schedule).

Expr

Fact

Facts are the left-hand side of a Command::Rule. They represent a part of a database query. Facts can be expressions or equality constraints between expressions.

Literal

NCommand

NormAction

NormExpr

NormFact

NormSchedule

Schedule

Type Definitions

CommandId