Enum egglog::ast::GenericNCommand

pub enum GenericNCommand<Head, Leaf, Ann>where
    Ann: Clone + Default,
    Head: Clone + Display,
    Leaf: Clone + PartialEq + Eq + Display + Hash,{

    SetOption {
        name: Symbol,
        value: GenericExpr<Head, Leaf, Ann>,
    },
    Sort(Symbol, Option<(Symbol, Vec<GenericExpr<Symbol, Symbol, Ann>>)>),
    Function(GenericFunctionDecl<Head, Leaf, Ann>),
    AddRuleset(Symbol),
    UnstableCombinedRuleset(Symbol, Vec<Symbol>),
    NormRule {
        name: Symbol,
        ruleset: Symbol,
        rule: GenericRule<Head, Leaf, Ann>,
    },
    CoreAction(GenericAction<Head, Leaf, Ann>),
    RunSchedule(GenericSchedule<Head, Leaf, Ann>),
    PrintOverallStatistics,
    Check(Vec<GenericFact<Head, Leaf, Ann>>),
    CheckProof,
    PrintTable(Symbol, usize),
    PrintSize(Option<Symbol>),
    Output {
        file: String,
        exprs: Vec<GenericExpr<Head, Leaf, Ann>>,
    },
    Push(usize),
    Pop(usize),
    Fail(Box<GenericNCommand<Head, Leaf, Ann>>),
    Input {
        name: Symbol,
        file: String,
    },

}

A NCommand is a desugared Command, where syntactic sugars like Command::Datatype, Command::Declare, and Command::Rewrite are eliminated. Most of the heavy lifting in egglog is done over NCommands.

GenericNCommand is a generalization of NCommand, like how GenericCommand is a generalization of Command, allowing annotations over Head and Leaf.

TODO: The name “NCommand” used to denote normalized command, but this meaning is obsolete. A future PR should rename this type to something like “DCommand”.

Variants

SetOption

Fields

name: Symbol

value: GenericExpr<Head, Leaf, Ann>

Sort(Symbol, Option<(Symbol, Vec<GenericExpr<Symbol, Symbol, Ann>>)>)

Function(GenericFunctionDecl<Head, Leaf, Ann>)

AddRuleset(Symbol)

UnstableCombinedRuleset(Symbol, Vec<Symbol>)

NormRule

Fields

name: Symbol

ruleset: Symbol

rule: GenericRule<Head, Leaf, Ann>

CoreAction(GenericAction<Head, Leaf, Ann>)

RunSchedule(GenericSchedule<Head, Leaf, Ann>)

PrintOverallStatistics

Check(Vec<GenericFact<Head, Leaf, Ann>>)

CheckProof

PrintTable(Symbol, usize)

PrintSize(Option<Symbol>)

Output

Fields

file: String

exprs: Vec<GenericExpr<Head, Leaf, Ann>>

Push(usize)

Pop(usize)

Fail(Box<GenericNCommand<Head, Leaf, Ann>>)

Input

Fields

name: Symbol

file: String

Implementations

impl<Head, Leaf> GenericNCommand<Head, Leaf, ()>where Head: Clone + Display, Leaf: Clone + PartialEq + Eq + Display + Hash,

pub fn to_command(&self) -> GenericCommand<Head, Leaf>

pub fn visit_exprs( self, f: &mut impl FnMut(GenericExpr<Head, Leaf, ()>) -> GenericExpr<Head, Leaf, ()> ) -> Self