Struct egglog::ast::Symbol

pub struct Symbol(/* private fields */);

A interned string in the global symbol table.

This requires the global feature on the crate.

GlobalSymbol is a wrapper around [Symbol] that knows to refer to a built-in, global [SymbolTable]. Strings into the global table are never freed.

This enables a lot of convience methods and trait implementations over GlobalSymbol (see below). In particular, you can convert it to &'static str, convert From and Into a &str, and de/serialize using serde if the serde feature is enabled.

Implementations

impl GlobalSymbol

pub fn new(s: impl AsRef<str>) -> GlobalSymbol

Intern a string into the global symbol table.

pub fn as_str(&self) -> &'static str

Convert this symbol into the string in the static, global symbol table.