egglog::ast

Type Alias FunctionDecl

pub type FunctionDecl = GenericFunctionDecl<String, String>;

Aliased Type

struct FunctionDecl {
    pub name: String,
    pub subtype: FunctionSubtype,
    pub schema: Schema,
    pub merge: Option<GenericExpr<String, String>>,
    pub cost: Option<u64>,
    pub unextractable: bool,
    pub let_binding: bool,
    pub span: Span,
}

Fields
name: Stringsubtype: FunctionSubtypeschema: Schemamerge: Option<GenericExpr<String, String>>cost: Option<u64>unextractable: boollet_binding: bool

Globals are desugared to functions, with this flag set to true. This is used by visualization to handle globals differently.

span: Span
Implementations

impl FunctionDecl

pub fn function( span: Span, name: String, schema: Schema, merge: Option<GenericExpr<String, String>>, ) -> Self

Constructs a function

pub fn constructor( span: Span, name: String, schema: Schema, cost: Option<DefaultCost>, unextractable: bool, ) -> Self

Constructs a constructor

pub fn relation(span: Span, name: String, input: Vec<String>) -> Self

Constructs a relation