egglog::ast

Type Alias ResolvedCommand

pub type ResolvedCommand = GenericCommand<ResolvedCall, ResolvedVar>;

Aliased Type

enum ResolvedCommand {

    Sort {
        span: Span,
        name: String,
        presort_and_args: Option<(String, Vec<GenericExpr<String, String>>)>,
        uf: Option<String>,
        unionable: bool,
    },
    Datatype {
        span: Span,
        name: String,
        variants: Vec<Variant>,
    },
    Datatypes {
        span: Span,
        datatypes: Vec<(Span, String, Subdatatypes)>,
    },
    Constructor {
        span: Span,
        name: String,
        schema: Schema,
        cost: Option<u64>,
        unextractable: bool,
        hidden: bool,
        let_binding: bool,
        term_constructor: Option<String>,
    },
    Relation {
        span: Span,
        name: String,
        inputs: Vec<String>,
    },
    Function {
        span: Span,
        name: String,
        schema: Schema,
        merge: Option<GenericExpr<ResolvedCall, ResolvedVar>>,
        hidden: bool,
        let_binding: bool,
    },
    AddRuleset(Span, String),
    UnstableCombinedRuleset(Span, String, Vec<String>),
    Rule {
        rule: GenericRule<ResolvedCall, ResolvedVar>,
    },
    Rewrite(String, GenericRewrite<ResolvedCall, ResolvedVar>, bool),
    BiRewrite(String, GenericRewrite<ResolvedCall, ResolvedVar>),
    Action(GenericAction<ResolvedCall, ResolvedVar>),
    Extract(Span, GenericExpr<ResolvedCall, ResolvedVar>, GenericExpr<ResolvedCall, ResolvedVar>),
    RunSchedule(GenericSchedule<ResolvedCall, ResolvedVar>),
    PrintOverallStatistics(Span, Option<String>),
    Check(Span, Vec<GenericFact<ResolvedCall, ResolvedVar>>),
    Prove(Span, Vec<GenericFact<ResolvedCall, ResolvedVar>>),
    ProveExists(Span, ResolvedCall),
    PrintFunction(Span, String, Option<usize>, Option<String>, PrintFunctionMode),
    PrintSize(Span, Option<String>),
    Input {
        span: Span,
        name: String,
        file: String,
    },
    Output {
        span: Span,
        file: String,
        exprs: Vec<GenericExpr<ResolvedCall, ResolvedVar>>,
    },
    Push(usize),
    Pop(Span, usize),
    Fail(Span, Box<GenericCommand<ResolvedCall, ResolvedVar>>),
    Include(Span, String),
    UserDefined(Span, String, Vec<GenericExpr<String, String>>),

}

Variants

Sort

Create a new user-defined sort, which can then be used in new Command::Function declarations. The Command::Datatype command desugars directly to this command, with one Command::Function per constructor. The main use of this command (as opposed to using Command::Datatype) is for forward-declaring a sort for mutually-recursive datatypes.

It can also be used to create a container sort. For example, here’s how to make a sort for vectors of some user-defined sort Math:

(sort MathVec (Vec Math))

Now MathVec can be used as an input or output sort.

Fields

span: Span

name: String

presort_and_args: Option<(String, Vec<GenericExpr<String, String>>)>

uf: Option<String>

The name of the union-find function for this sort. Used in term encoding to canonicalize values during extraction.

unionable: bool

Whether values of this sort can be unioned. Defaults to true for user-defined sorts. Set to false for relations and term tables that should not allow union.

Datatype

Egglog supports three types of functions

A constructor models an egg-style user-defined datatype It can only be defined through the datatype/datatype* command or the constructor command

A relation models a datalog-style mathematical relation It can only be defined through the relation command

A custom function is a dictionary It can only be defined through the function command

The datatype command declares a user-defined datatype. Datatypes can be unioned with Action::Union either at the top level or in the actions of a rule. This makes them equal in the implicit, global equality relation.

Example:

(datatype Math
  (Num i64)
  (Var String)
  (Add Math Math)
  (Mul Math Math))

defines a simple Math datatype with variants for numbers, named variables, addition and multiplication.

Datatypes desugar directly to a Command::Sort and a Command::Constructor for each constructor. The code above becomes:

(sort Math)
(constructor Num (i64) Math)
(constructor Var (String) Math)
(constructor Add (Math Math) Math)
(constructor Mul (Math Math) Math)

Datatypes are also known as algebraic data types, tagged unions and sum types.

Fields

span: Span

name: String

variants: Vec<Variant>

Datatypes

Fields

span: Span

datatypes: Vec<(Span, String, Subdatatypes)>

Constructor

The constructor command defines a new constructor for a user-defined datatype Example:

(constructor Add (i64 i64) Math)

A constructor can be :unextractable, in which case extraction skip this constructor entirely.

(constructor UnextractableNode (i64) Math :unextractable)

A constructor can also have a cost for extraction, which is used in the cost model for extraction. It defaults to 1 if not specified.

(constructor ExpensiveNode (i64) Math :cost 10)

Fields

span: Span

name: String

schema: Schema

cost: Option<u64>

unextractable: bool

hidden: bool

Internal-hidden constructors are excluded from print-size output and extraction. Used for internal tables generated by proof production.

let_binding: bool

Internal-let constructors are let bindings, excluded from print-size output. Used for global let bindings that are converted to constructors.

term_constructor: Option<String>

For view tables in proof encoding: the constructor to use for building terms from the first n-1 children during extraction.

Relation

The relation command declares a named relation Example:

(relation path (i64 i64))
(relation edge (i64 i64))

Fields

span: Span

name: String

inputs: Vec<String>

Function

The function command declare an egglog custom function, which is a database table with a a functional dependency (also called a primary key) on its inputs to one output.

(function <name:Ident> <schema:Schema> <cost:Cost>
       (:on_merge <List<Action>>)?
       (:merge <Expr>)?)

A function can have a cost for extraction.

Finally, it can have a merge and on_merge, which are triggered when the function dependency is violated. In this case, the merge expression determines which of the two outputs for the same input is used. The on_merge actions are run after the merge expression is evaluated.

Note that the :merge expression must be monotonic for the behavior of the egglog program to be consistent and defined. In other words, the merge function must define a lattice on the output of the function. If values are merged in different orders, they should still result in the same output. If the merge expression is not monotonic, the behavior can vary as actions may be applied more than once with different results.

(function LowerBound (Math) i64 :merge (max old new))

Specifically, a custom function can also have an EqSort output type:

(function Add (i64 i64) Math)

All functions can be set with Action::Set.

Output of a function, if being the EqSort type, can be unioned with Action::Union with another datatype of the same sort.

Fields

span: Span

name: String

schema: Schema

merge: Option<GenericExpr<ResolvedCall, ResolvedVar>>

hidden: bool

let_binding: bool

AddRuleset(Span, String)

Using the ruleset command, defines a new ruleset that can be added to in Command::Rules. Rulesets are used to group rules together so that they can be run together in a Schedule.

Example: Ruleset allows users to define a ruleset- a set of rules

(ruleset myrules)
(rule ((edge x y))
      ((path x y))
      :ruleset myrules)
(run myrules 2)

UnstableCombinedRuleset(Span, String, Vec<String>)

Using the combined-ruleset command, construct another ruleset which runs all the rules in the given rulesets. This is useful for running multiple rulesets together. The combined ruleset also inherits any rules added to the individual rulesets after the combined ruleset is declared.

Example:

(ruleset myrules1)
(rule ((edge x y))
      ((path x y))
     :ruleset myrules1)
(ruleset myrules2)
(rule ((path x y) (edge y z))
      ((path x z))
      :ruleset myrules2)
(combined-ruleset myrules-combined myrules1 myrules2)

Rule

(rule <body:List<Fact>> <head:List<Action>>)

defines an egglog rule. The rule matches a list of facts with respect to the global database, and runs the list of actions for each match. The matches are done modulo equality, meaning equal datatypes in the database are considered equal.

Example:

(rule ((edge x y))
      ((path x y)))

(rule ((path x y) (edge y z))
      ((path x z)))

Fields

rule: GenericRule<ResolvedCall, ResolvedVar>

Rewrite(String, GenericRewrite<ResolvedCall, ResolvedVar>, bool)

rewrite is syntactic sugar for a specific form of rule which simply unions the left and right hand sides.

Example:

(rewrite (Add a b)
         (Add b a))

Desugars to:

(rule ((= lhs (Add a b)))
      ((union lhs (Add b a))))

Additionally, additional facts can be specified using a :when clause. For example, the same rule can be run only when a is zero:

(rewrite (Add a b)
         (Add b a)
         :when ((= a (Num 0)))

Add the :subsume flag to cause the left hand side to be subsumed after matching, which means it can no longer be matched in a rule, but can still be checked against (See Change for more details.)

(rewrite (Mul a 2) (bitshift-left a 1) :subsume)

Desugars to:

(rule ((= lhs (Mul a 2)))
      ((union lhs (bitshift-left a 1))
       (subsume (Mul a 2))))

BiRewrite(String, GenericRewrite<ResolvedCall, ResolvedVar>)

Similar to Command::Rewrite, but generates two rules, one for each direction.

Example:

(bi-rewrite (Mul (Var x) (Num 0))
            (Var x))

Becomes:

(rule ((= lhs (Mul (Var x) (Num 0))))
      ((union lhs (Var x))))
(rule ((= lhs (Var x)))
      ((union lhs (Mul (Var x) (Num 0)))))

Action(GenericAction<ResolvedCall, ResolvedVar>)

Perform an Action on the global database (see documentation for Action for more details). Example:

(let xplusone (Add (Var "x") (Num 1)))

Extract(Span, GenericExpr<ResolvedCall, ResolvedVar>, GenericExpr<ResolvedCall, ResolvedVar>)

extract a datatype from the egraph, choosing the smallest representative. By default, each constructor costs 1 to extract (common subexpressions are not shared in the cost model).

RunSchedule(GenericSchedule<ResolvedCall, ResolvedVar>)

Runs a Schedule, which specifies rulesets and the number of times to run them.

Example:

(run-schedule
    (saturate my-ruleset-1)
    (run my-ruleset-2 4))

Runs my-ruleset-1 until saturation, then runs my-ruleset-2 four times.

See Schedule for more details.

PrintOverallStatistics(Span, Option<String>)

Print runtime statistics about rules and rulesets so far.

Check(Span, Vec<GenericFact<ResolvedCall, ResolvedVar>>)

The check command checks that the given facts match at least once in the current database. The list of facts is matched in the same way a Command::Rule is matched.

Example:

(check (= (+ 1 2) 3))
(check (<= 0 3) (>= 3 0))
(fail (check (= 1 2)))

prints

[INFO ] Checked.
[INFO ] Checked.
[ERROR] Check failed
[INFO ] Command failed as expected.

Prove(Span, Vec<GenericFact<ResolvedCall, ResolvedVar>>)

ProveExists(Span, ResolvedCall)

PrintFunction(Span, String, Option<usize>, Option<String>, PrintFunctionMode)

Print out rows of a given function, extracting each of the elements of the function. Example:

(print-function Add 20)

prints the first 20 rows of the Add function.

(print-function Add)

prints all rows of the Add function.

(print-function Add :file "add.csv")

prints all rows of the Add function to a CSV file.

PrintSize(Span, Option<String>)

Print out the number of rows in a function or all functions.

Input

Input a CSV file directly into a function.

Fields

span: Span

name: String

file: String

Output

Extract and output a set of expressions to a file.

Fields

span: Span

file: String

exprs: Vec<GenericExpr<ResolvedCall, ResolvedVar>>

Push(usize)

push the current egraph n times so that it is saved. Later, the current database and rules can be restored using pop.

Pop(Span, usize)

pop the current egraph, restoring the previous one. The argument specifies how many egraphs to pop.

Fail(Span, Box<GenericCommand<ResolvedCall, ResolvedVar>>)

Assert that a command fails with an error.

Include(Span, String)

Include another egglog file directly as text and run it.

UserDefined(Span, String, Vec<GenericExpr<String, String>>)

User-defined command.