egglog::ast

Type Alias ResolvedExpr

pub type ResolvedExpr = GenericExpr<ResolvedCall, ResolvedVar>;

A generated expression is an expression that is generated by the system and does not have annotations.

Aliased Type

enum ResolvedExpr {
    Var(Span, ResolvedVar),
    Call(Span, ResolvedCall, Vec<GenericExpr<ResolvedCall, ResolvedVar>>),
    Lit(Span, Literal),
}

Variants

Var(Span, ResolvedVar)

Call(Span, ResolvedCall, Vec<GenericExpr<ResolvedCall, ResolvedVar>>)

Lit(Span, Literal)