egglog::extract

Trait Cost

pub trait Cost {
    // Required methods
    fn identity() -> Self;
    fn unit() -> Self;
    fn combine(self, other: &Self) -> Self;
}

Requirements for a type to be usable as a cost by a CostModel.

Required Methods

fn identity() -> Self

An identity element, usually zero.

fn unit() -> Self

The default cost for a node with no children, usually one.

fn combine(self, other: &Self) -> Self

A binary operation to combine costs, usually addition. This operation must NOT overflow or panic when given large values!

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.