egglog::prelude

Trait BaseSort

pub trait BaseSort:
    Any
    + Send
    + Sync
    + Debug {
    type Base: BaseValue;

    // Required methods
    fn name(&self) -> &str;
    fn reconstruct_termdag(
        &self,
        _: &BaseValues,
        _: Value,
        _: &mut TermDag,
    ) -> TermId;

    // Provided methods
    fn register_primitives(&self, _eg: &mut EGraph) { ... }
    fn to_arcsort(self) -> ArcSort
       where Self: Sized { ... }
}

A “default” implementation of Sort for simple types which just want to put some data in the e-graph. If you implement this trait, do not implement Sort or ContainerSort. Use add_base_sortto register base sorts with theEGraph. See Sortfor documentation of the methods. Do not overrideto_arcsort`.

Required Associated Types

type Base: BaseValue

Required Methods

fn name(&self) -> &str

fn reconstruct_termdag( &self, _: &BaseValues, _: Value, _: &mut TermDag, ) -> TermId

Provided Methods

fn register_primitives(&self, _eg: &mut EGraph)

fn to_arcsort(self) -> ArcSort
where Self: Sized,