Module egglog::ast

Modules

desugar

parse

Structs

CorrespondingVar

Facts

GenericActions

GenericFunctionDecl

Represents the declaration of a function directly parsed from source syntax.

GenericRewrite

GenericRule

GenericRunConfig

Id

IdentSort

ResolvedVar

Schema

Symbol

A interned string in the global symbol table.

Variant

Enums

Change

Change a function entry.

GenericAction

GenericCommand

A Command is the top-level construct in egglog. It includes defining rules, declaring functions, adding to tables, and running rules (via a Schedule).

GenericExpr

GenericFact

Facts are the left-hand side of a Command::Rule. They represent a part of a database query. Facts can be expressions or equality constraints between expressions.

GenericNCommand

A NCommand is a desugared Command, where syntactic sugars like Command::Datatype, Command::Declare, and Command::Rewrite are eliminated. Most of the heavy lifting in egglog is done over NCommands.

GenericSchedule

Literal

Traits

ToSexp

Type Aliases

Action

Actions

Command

Expr

Fact

FunctionDecl

NCommand

Rule

RunConfig

Schedule

Subsume