egglog::ast

Function sanitize_internal_names

pub fn sanitize_internal_names<Head, Leaf>(
    program: &[GenericCommand<Head, Leaf>],
) -> Vec<GenericCommand<String, String>>
where
    Head: Clone + Display,
    Leaf: Clone + PartialEq + Eq + Display + Hash,

Sanitizes internal names so they do not contain any internal characters. This enables printing desugared egglog in a way that can be re-parsed.