Type Alias egglog::ast::FunctionDecl

pub type FunctionDecl = GenericFunctionDecl<Symbol, Symbol>;

Aliased Type

struct FunctionDecl {
    pub name: GlobalSymbol,
    pub subtype: FunctionSubtype,
    pub schema: Schema,
    pub merge: Option<GenericExpr<GlobalSymbol, GlobalSymbol>>,
    pub cost: Option<usize>,
    pub unextractable: bool,
    pub ignore_viz: bool,
    pub span: Span,
}

Fields
name: GlobalSymbolsubtype: FunctionSubtypeschema: Schemamerge: Option<GenericExpr<GlobalSymbol, GlobalSymbol>>cost: Option<usize>unextractable: boolignore_viz: bool

Globals are desugared to functions, with this flag set to true. This is used by visualization to handle globals differently.

span: Span
Implementations

impl FunctionDecl

pub fn function( span: Span, name: Symbol, schema: Schema, merge: Option<GenericExpr<Symbol, Symbol>> ) -> Self

pub fn constructor( span: Span, name: Symbol, schema: Schema, cost: Option<usize>, unextractable: bool ) -> Self

pub fn relation(span: Span, name: Symbol, input: Vec<Symbol>) -> Self