egglog::ast

Type Alias FunctionDecl

pub type FunctionDecl = GenericFunctionDecl<String, String>;

Aliased Type

struct FunctionDecl {
    pub name: String,
    pub subtype: FunctionSubtype,
    pub schema: Schema,
    pub resolved_schema: String,
    pub merge: Option<GenericExpr<String, String>>,
    pub cost: Option<u64>,
    pub unextractable: bool,
    pub internal_hidden: bool,
    pub internal_let: bool,
    pub span: Span,
    pub term_constructor: Option<String>,
}

Fields
name: Stringsubtype: FunctionSubtypeschema: Schema

Untyped schema

resolved_schema: String

Resolved schema after typechecking is stored here, otherwise “”.

merge: Option<GenericExpr<String, String>>cost: Option<u64>unextractable: boolinternal_hidden: bool

Hidden functions are excluded from print-size output. Used for internal tables generated by proof production.

internal_let: bool

Globals are desugared to functions, with this flag set to true. This is used by visualization to handle globals differently.

span: Spanterm_constructor: Option<String>

For view tables in proof encoding: the constructor to use for building terms from the first n-1 children during extraction.

Implementations

impl FunctionDecl

pub fn function( span: Span, name: String, schema: Schema, merge: Option<GenericExpr<String, String>>, ) -> Self

Constructs a function

pub fn constructor( span: Span, name: String, schema: Schema, cost: Option<DefaultCost>, unextractable: bool, hidden: bool, ) -> Self

Constructs a constructor