Trait egglog::sort::Sort

pub trait Sort: Any + Send + Sync + Debug {
    // Required methods
    fn name(&self) -> Symbol;
    fn as_arc_any(self: Arc<Self>) -> Arc<dyn Any + Send + Sync + 'static>;
    fn make_expr(&self, egraph: &EGraph, value: Value) -> (usize, Expr);

    // Provided methods
    fn is_eq_sort(&self) -> bool { ... }
    fn is_container_sort(&self) -> bool { ... }
    fn is_eq_container_sort(&self) -> bool { ... }
    fn foreach_tracked_values<'a>(
        &'a self,
        value: &'a Value,
        f: Box<dyn FnMut(Value) + 'a>
    ) { ... }
    fn canonicalize(&self, value: &mut Value, unionfind: &UnionFind) -> bool { ... }
    fn serialized_name(&self, _value: &Value) -> Symbol { ... }
    fn inner_values(&self, value: &Value) -> Vec<(ArcSort, Value)> { ... }
    fn register_primitives(self: Arc<Self>, info: &mut TypeInfo) { ... }
    fn extract_expr(
        &self,
        egraph: &EGraph,
        value: Value,
        _extractor: &Extractor<'_>,
        _termdag: &mut TermDag
    ) -> Option<(usize, Expr)> { ... }
}

Required Methods

fn name(&self) -> Symbol

fn as_arc_any(self: Arc<Self>) -> Arc<dyn Any + Send + Sync + 'static>

fn make_expr(&self, egraph: &EGraph, value: Value) -> (usize, Expr)

Extracting an expression (with smallest cost) out of a primitive value

Provided Methods

fn is_eq_sort(&self) -> bool

fn is_container_sort(&self) -> bool

fn is_eq_container_sort(&self) -> bool

fn foreach_tracked_values<'a>( &'a self, value: &'a Value, f: Box<dyn FnMut(Value) + 'a> )

fn canonicalize(&self, value: &mut Value, unionfind: &UnionFind) -> bool

fn serialized_name(&self, _value: &Value) -> Symbol

Return the serialized name of the sort

Only used for container sorts, which cannot be serialized with make_expr so need an explicit name

fn inner_values(&self, value: &Value) -> Vec<(ArcSort, Value)>

Return the inner values and sorts. Only eq_container_sort need to implement this method,

fn register_primitives(self: Arc<Self>, info: &mut TypeInfo)

fn extract_expr( &self, egraph: &EGraph, value: Value, _extractor: &Extractor<'_>, _termdag: &mut TermDag ) -> Option<(usize, Expr)>

For values like EqSort containers, to make/extract an expression from it requires an extractor. Moreover, the extraction may be unsuccessful if the extractor is not fully initialized.

The default behavior is to call make_expr