Struct egglog::EGraph

pub struct EGraph {
    pub functions: HashMap<Symbol, Function, BuildHasherDefault<FxHasher>>,
    pub run_mode: RunMode,
    pub test_proofs: bool,
    pub match_limit: usize,
    pub node_limit: usize,
    pub fact_directory: Option<PathBuf>,
    pub seminaive: bool,
    /* private fields */
}

Fields
functions: HashMap<Symbol, Function, BuildHasherDefault<FxHasher>>run_mode: RunModetest_proofs: boolmatch_limit: usizenode_limit: usizefact_directory: Option<PathBuf>seminaive: bool
Implementations

impl EGraph

pub fn extract( &self, value: Value, termdag: &mut TermDag, arcsort: &ArcSort ) -> (usize, Term)

This example uses EGraph::extract to extract a term. The example is trivial, as there is only a single variant of the expression in the egraph.

use egglog::{EGraph, TermDag};
let mut egraph = EGraph::default();
egraph
    .parse_and_run_program(
        "(datatype Op (Add i64 i64))
         (let expr (Add 1 1))",
    )
    .unwrap();
let mut termdag = TermDag::default();
let (sort, value) = egraph
    .eval_expr(&egglog::ast::Expr::Var((), "expr".into()))
    .unwrap();
let (_, extracted) = egraph.extract(value, &mut termdag, &sort);
assert_eq!(termdag.to_string(&extracted), "(Add 1 1)");

pub fn extract_variants( &mut self, value: Value, limit: usize, termdag: &mut TermDag ) -> Vec<Term>

impl EGraph

pub fn serialize(&self, config: SerializeConfig) -> EGraph

Serialize the egraph into a format that can be read by the egraph-serialize crate.

There are multiple different semantically valid ways to do this. This is how this implementation does it:

For node costs:

Primitives: 1.0

Function without costs: 1.0

Function with costs: the cost

For node IDs:

Functions: Function name + hash of input values

Args which are eq sorts: Choose one ID from the e-class, distribute roughly evenly.

Args and outputs values which are primitives: Sort name + hash of value Notes: If split_primitive_returns is true, then each output value will be the function node id + -output

For e-classes:

Eq sorts: Use the canonical ID of the e-class

Primitives: Use the node ID

This is to achieve the following properties:

Equivalent primitive values will show up once in the e-graph.

Functions which return primitive values will be added to the e-class of that value.

Nodes will have consistant IDs throughout execution of e-graph (used for animating changes in the visualization)

Edges in the visualization will be well distributed (used for animating changes in the visualization) (Note that this will be changed in <https://github.com/egraphs-good/egglog/pull/158> so that edges point to exact nodes instead of looking up the e-class)

impl EGraph

pub fn enable_terms_encoding(&mut self)

Use the rust backend implimentation of eqsat, including a rust implementation of the union-find data structure and the rust implementation of the rebuilding algorithm (maintains congruence closure).

pub fn is_interactive_mode(&self) -> bool

pub fn push(&mut self)

pub fn pop(&mut self) -> Result<(), Error>

Pop the current egraph off the stack, replacing it with the previously pushed egraph. It preserves the run report and messages from the popped egraph.

pub fn union(&mut self, id1: Id, id2: Id, sort: Symbol) -> Id

pub fn find(&self, value: Value) -> Value

find the leader value for a particular eclass

pub fn rebuild_nofail(&mut self) -> usize

pub fn rebuild(&mut self) -> Result<usize, Error>

pub fn eval_lit(&self, lit: &Literal) -> Value

pub fn function_to_dag( &mut self, sym: Symbol, n: usize ) -> Result<(Vec<(Term, Term)>, TermDag), Error>

pub fn print_function(&mut self, sym: Symbol, n: usize) -> Result<(), Error>

pub fn print_size(&mut self, sym: Option<Symbol>) -> Result<(), Error>

pub fn extract_value(&self, value: Value) -> (TermDag, Term)

Extract a value to a TermDag and Term in the TermDag. See also extract_value_to_string for convenience.

pub fn extract_value_to_string(&self, value: Value) -> String

Extract a value to a string for printing. See also extract_value for more control.

pub fn eval_expr(&mut self, expr: &Expr) -> Result<(ArcSort, Value), Error>

pub fn clear(&mut self)

pub fn set_reserved_symbol(&mut self, sym: Symbol)

pub fn run_program( &mut self, program: Vec<Command> ) -> Result<Vec<String>, Error>

Run a program, represented as an AST. Return a list of messages.

pub fn parse_program(&self, input: &str) -> Result<Vec<Command>, Error>

pub fn parse_and_run_program( &mut self, input: &str ) -> Result<Vec<String>, Error>

pub fn num_tuples(&self) -> usize

pub fn get_sort_by<S: Sort + Send + Sync>( &self, pred: impl Fn(&Arc<S>) -> bool ) -> Option<Arc<S>>

Returns the first sort that satisfies the type and predicate if there’s one. Otherwise returns none.

pub fn get_sort<S: Sort + Send + Sync>(&self) -> Option<Arc<S>>

Returns a sort based on the type

pub fn add_arcsort(&mut self, arcsort: ArcSort) -> Result<(), TypeError>

Add a user-defined sort

pub fn add_primitive(&mut self, prim: impl Into<Primitive>)

Add a user-defined primitive

pub fn get_extract_report(&self) -> &Option<ExtractReport>

Gets the last extract report and returns it, if the last command saved it.

pub fn get_run_report(&self) -> &Option<RunReport>

Gets the last run report and returns it, if the last command saved it.

pub fn get_overall_run_report(&self) -> &RunReport

Gets the overall run report and returns it.

pub fn serialize_for_graphviz( &self, split_primitive_outputs: bool, max_functions: usize, max_calls_per_function: usize ) -> EGraph

Serializes the egraph for export to graphviz.

This will limit the total number of nodes so that visualization does not blow up.