Struct egglog::EGraph

pub struct EGraph {
    pub parser: Parser,
    pub functions: IndexMap<Symbol, Function>,
    pub run_mode: RunMode,
    pub fact_directory: Option<PathBuf>,
    pub seminaive: bool,
    /* private fields */
}

Fields
parser: Parserfunctions: IndexMap<Symbol, Function>run_mode: RunModefact_directory: Option<PathBuf>seminaive: bool
Implementations

impl EGraph

pub fn repl(&mut self) -> Result<()>

pub fn repl_with<R, W>(&mut self, input: R, output: W) -> Result<()>
where R: Read, W: Write,

impl EGraph

pub fn extract( &self, value: Value, termdag: &mut TermDag, arcsort: &ArcSort ) -> Result<(Cost, Term), Error>

This example uses EGraph::extract to extract a term. The example is trivial, as there is only a single variant of the expression in the egraph.

use egglog::{EGraph, TermDag};
let mut egraph = EGraph::default();
egraph
    .parse_and_run_program(
        None,
        "(datatype Op (Add i64 i64))
         (let expr (Add 1 1))",
    )
    .unwrap();
let mut termdag = TermDag::default();
let (sort, value) = egraph.eval_expr(&egglog::var!("expr")).unwrap();
let (_, extracted) = egraph.extract(value, &mut termdag, &sort).unwrap();
assert_eq!(termdag.to_string(&extracted), "(Add 1 1)");

pub fn extract_variants( &mut self, sort: &ArcSort, value: Value, limit: usize, termdag: &mut TermDag ) -> Vec<Term>

impl EGraph

pub fn serialize(&self, config: SerializeConfig) -> EGraph

Serialize the egraph into a format that can be read by the egraph-serialize crate.

There are multiple different semantically valid ways to do this. This is how this implementation does it:

For node costs:

Primitives: 1.0

Function without costs: 1.0

Function with costs: the cost

Omitted nodes: infinite

For node IDs:

Functions: Function name + hash of input values

Args which are eq sorts: Choose one ID from the e-class, distribute roughly evenly.

Args and outputs values which are primitives: Sort name + hash of value

For e-classes IDs:

tag and value of canonicalized value

This is to achieve the following properties:

Equivalent primitive values will show up once in the e-graph.

Functions which return primitive values will be added to the e-class of that value.

Nodes will have consistant IDs throughout execution of e-graph (used for animating changes in the visualization)

Edges in the visualization will be well distributed (used for animating changes in the visualization) (Note that this will be changed in <https://github.com/egraphs-good/egglog/pull/158> so that edges point to exact nodes instead of looking up the e-class)

pub fn value_to_class_id(&self, sort: &ArcSort, value: &Value) -> ClassId

Gets the serialized class ID for a value.

pub fn class_id_to_value(&self, eclass_id: &ClassId) -> Value

Gets the value for a serialized class ID.

pub fn to_node_id(&self, sort: Option<&ArcSort>, node: SerializedNode) -> NodeId

Gets the serialized node ID for the primitive, omitted, or function value.

pub fn from_node_id(&self, node_id: &NodeId) -> SerializedNode

Gets the serialized node for the node ID.

impl EGraph

pub fn is_interactive_mode(&self) -> bool

pub fn push(&mut self)

pub fn disable_messages(&mut self)

Disable saving messages to be printed to the user and remove any saved messages.

When messages are disabled the vec of messages returned by evaluating commands will always be empty.

pub fn enable_messages(&mut self)

Enable saving messages to be printed to the user.

pub fn messages_enabled(&self) -> bool

Whether messages are enabled.

pub fn pop(&mut self) -> Result<(), Error>

Pop the current egraph off the stack, replacing it with the previously pushed egraph. It preserves the run report and messages from the popped egraph.

pub fn union(&mut self, id1: u64, id2: u64, sort: Symbol) -> u64

pub fn find(&self, sort: &ArcSort, value: Value) -> Value

find the leader value for a particular eclass

pub fn rebuild_nofail(&mut self) -> usize

pub fn rebuild(&mut self) -> Result<usize, Error>

pub fn eval_lit(&self, lit: &Literal) -> Value

pub fn function_to_dag( &mut self, sym: Symbol, n: usize ) -> Result<(Vec<(Term, Term)>, TermDag), Error>

pub fn print_function(&mut self, sym: Symbol, n: usize) -> Result<(), Error>

pub fn print_size(&mut self, sym: Option<Symbol>) -> Result<(), Error>

pub fn extract_value( &self, sort: &ArcSort, value: Value ) -> Result<(TermDag, Term), Error>

Extract a value to a TermDag and Term in the TermDag. Note that the TermDag may contain a superset of the nodes in the Term. See also extract_value_to_string for convenience.

pub fn extract_value_to_string( &self, sort: &ArcSort, value: Value ) -> Result<String, Error>

Extract a value to a string for printing. See also extract_value for more control.

pub fn eval_expr(&mut self, expr: &Expr) -> Result<(ArcSort, Value), Error>

pub fn clear(&mut self)

pub fn set_reserved_symbol(&mut self, sym: Symbol)

pub fn run_program( &mut self, program: Vec<Command> ) -> Result<Vec<String>, Error>

Run a program, represented as an AST. Return a list of messages.

pub fn parse_and_run_program( &mut self, filename: Option<String>, input: &str ) -> Result<Vec<String>, Error>

Takes a source program input, parses it, runs it, and returns a list of messages.

filename is an optional argument to indicate the source of the program for error reporting. If filename is None, a default name will be used.

pub fn num_tuples(&self) -> usize

pub fn get_sort<S: Sort + Send + Sync>(&self) -> Option<Arc<S>>

Returns a sort based on the type

pub fn get_sort_by<S: Sort + Send + Sync>( &self, pred: impl Fn(&Arc<S>) -> bool ) -> Option<Arc<S>>

Returns the first sort that satisfies the type and predicate if there’s one. Otherwise returns none.

pub fn add_arcsort( &mut self, arcsort: ArcSort, span: Span ) -> Result<(), TypeError>

Add a user-defined sort

pub fn add_primitive(&mut self, prim: impl Into<Primitive>)

Add a user-defined primitive

pub fn get_extract_report(&self) -> &Option<ExtractReport>

Gets the last extract report and returns it, if the last command saved it.

pub fn get_run_report(&self) -> &Option<RunReport>

Gets the last run report and returns it, if the last command saved it.

pub fn get_overall_run_report(&self) -> &RunReport

Gets the overall run report and returns it.