Struct egglog::EGraph

pub struct EGraph {
    pub test_proofs: bool,
    pub match_limit: usize,
    pub node_limit: usize,
    pub fact_directory: Option<PathBuf>,
    pub seminaive: bool,
    pub global_bindings: HashMap<Symbol, (ArcSort, Value, u32), BuildHasherDefault<FxHasher>>,
    /* private fields */
}

Fields
test_proofs: boolmatch_limit: usizenode_limit: usizefact_directory: Option<PathBuf>seminaive: boolglobal_bindings: HashMap<Symbol, (ArcSort, Value, u32), BuildHasherDefault<FxHasher>>
Implementations

impl EGraph

pub fn extract( &self, value: Value, termdag: &mut TermDag, arcsort: &ArcSort ) -> (usize, Term)

This example uses EGraph::extract to extract a term. The example is trivial, as there is only a single variant of the expression in the egraph.

use egglog::{EGraph, TermDag};
let mut egraph = EGraph::default();
egraph
    .parse_and_run_program(
        "(datatype Op (Add i64 i64))
         (let expr (Add 1 1))",
    )
    .unwrap();
let mut termdag = TermDag::default();
let (sort, value) = egraph
    .eval_expr(&egglog::ast::Expr::Var("expr".into()), None, true)
    .unwrap();
let (_, extracted) = egraph.extract(value, &mut termdag, &sort);
assert_eq!(termdag.to_string(&extracted), "(Add 1 1)");

pub fn extract_variants( &mut self, value: Value, limit: usize, termdag: &mut TermDag ) -> Vec<Term>

impl EGraph

pub fn serialize(&self, config: SerializeConfig) -> EGraph

Serialize the egraph into a format that can be read by the egraph-serialize crate.

There are multiple different semantically valid ways to do this. This is how this implementation does it:

For node costs:

Primitives: 1.0

Function without costs: 1.0

Function with costs: the cost

For node IDs:

Functions: Function name + hash of input values

Args which are eq sorts: Choose one ID from the e-class, distribute roughly evenly.

Args and outputs values which are primitives: Sort name + hash of value Notes: If split_primitive_returns is true, then each output value will be the function node id + -output

For e-classes:

Eq sorts: Use the canonical ID of the e-class

Primitives: Use the node ID

This is to achieve the following properties:

Equivalent primitive values will show up once in the e-graph.

Functions which return primitive values will be added to the e-class of that value.

Nodes will have consistant IDs throughout execution of e-graph (used for animating changes in the visualization)

Edges in the visualization will be well distributed (used for animating changes in the visualization) (Note that this will be changed in <https://github.com/egraphs-good/egglog/pull/158> so that edges point to exact nodes instead of looking up the e-class)

impl EGraph

pub fn compile_actions( &self, types: &IndexMap<Symbol, ArcSort>, actions: &[Action] ) -> Result<Program, Vec<TypeError>>

pub fn compile_expr( &self, types: &IndexMap<Symbol, ArcSort>, expr: &Expr, expected_type: Option<ArcSort> ) -> Result<(ArcSort, Program), Vec<TypeError>>

pub fn run_actions( &mut self, stack: &mut Vec<Value>, subst: &[Value], program: &Program, make_defaults: bool ) -> Result<(), Error>

impl EGraph

pub fn enable_terms_encoding(&mut self)

Use the rust backend implimentation of eqsat, including a rust implementation of the union-find data structure and the rust implementation of the rebuilding algorithm (maintains congruence closure).

pub fn is_interactive_mode(&self) -> bool

pub fn push(&mut self)

pub fn pop(&mut self) -> Result<(), Error>

Pop the current egraph off the stack, replacing it with the previously pushed egraph. It preserves the run report and messages from the popped egraph.

pub fn union(&mut self, id1: Id, id2: Id, sort: Symbol) -> Id

pub fn find(&self, value: Value) -> Value

find the leader value for a particular eclass

pub fn rebuild_nofail(&mut self) -> usize

pub fn rebuild(&mut self) -> Result<usize, Error>

pub fn declare_function(&mut self, decl: &NormFunctionDecl) -> Result<(), Error>

pub fn eval_lit(&self, lit: &Literal) -> Value

pub fn function_to_dag( &mut self, sym: Symbol, n: usize ) -> Result<(Vec<(Term, Term)>, TermDag), Error>

pub fn print_function(&mut self, sym: Symbol, n: usize) -> Result<(), Error>

pub fn print_size(&mut self, sym: Option<Symbol>) -> Result<(), Error>

pub fn run_schedule(&mut self, sched: &NormSchedule) -> RunReport

pub fn extract_value(&self, value: Value) -> (TermDag, Term)

Extract a value to a TermDag and Term in the TermDag. See also extract_value_to_string for convenience.

pub fn extract_value_to_string(&self, value: Value) -> String

Extract a value to a string for printing. See also extract_value for more control.

pub fn run_rules(&mut self, config: &NormRunConfig) -> RunReport

pub fn add_rule(&mut self, rule: Rule, ruleset: Symbol) -> Result<Symbol, Error>

pub fn eval_actions(&mut self, actions: &[Action]) -> Result<(), Error>

pub fn eval_expr( &mut self, expr: &Expr, expected_type: Option<ArcSort>, make_defaults: bool ) -> Result<(ArcSort, Value), Error>

pub fn set_option(&mut self, name: &str, value: Expr)

pub fn clear(&mut self)

pub fn process_commands( &mut self, program: Vec<Command>, stop: CompilerPassStop ) -> Result<Vec<NormCommand>, Error>

pub fn set_underscores_for_desugaring(&mut self, underscores: usize)

pub fn run_program( &mut self, program: Vec<Command> ) -> Result<Vec<String>, Error>

pub fn parse_program(&self, input: &str) -> Result<Vec<Command>, Error>

pub fn parse_and_run_program( &mut self, input: &str ) -> Result<Vec<String>, Error>

pub fn num_tuples(&self) -> usize

pub fn add_arcsort(&mut self, arcsort: ArcSort) -> Result<(), TypeError>

pub fn get_extract_report(&self) -> &Option<ExtractReport>

Gets the last extract report and returns it, if the last command saved it.

pub fn get_run_report(&self) -> &Option<RunReport>

Gets the last run report and returns it, if the last command saved it.

pub fn get_overall_run_report(&self) -> &RunReport

Gets the overall run report and returns it.

pub fn serialize_for_graphviz(&self, split_primitive_outputs: bool) -> EGraph

Serializes the egraph for export to graphviz.